SPA is distributed between different modules
