+++ view/js/spa-router.js
@@ -16,7 +16,26 @@
  * To install Acorn: composer require npm-asset/acorn:^8.15.0
  */
 
-import * as acorn from '/view/asset/acorn/dist/acorn.mjs';
+import {
+  checkForDOMReadyPatterns,
+  classifyScriptContent,
+  promoteToGlobal,
+  classifyScript,
+  executeScripts
+} from '/view/js/spa-script-runtime.js';
+import {
+  showTimeoutModal,
+  dismissTimeoutModal,
+  cleanupTooltips
+} from '/view/js/spa-ui-helpers.js';
+import { createNavigationAdapter } from '/view/js/spa-navigation-adapter.js';
+import {
+  triggerSPADocumentReady,
+  triggerSPAWindowLoad,
+  reinitializeDynamicContent
+} from '/view/js/spa-lifecycle.js';
+import { createDomSwapPipeline } from '/view/js/spa-dom-swap.js';
+import { createContentLoader } from '/view/js/spa-content-loader.js';
 
 // ============================================
 // FEATURE DETECTION
@@ -24,24 +43,10 @@ import * as acorn from '/view/asset/acorn/dist/acorn.mjs';
 
 const supportsSPA = window.history && window.history.pushState && window.fetch;
 
-function normalizeScriptSource(src) {
-  if (!src) {
-    return null;
-  }
-
-  try {
-    const baseUrl = document.baseURI || (window.location.origin + '/');
-    return new URL(src, baseUrl).toString();
-  } catch (e) {
-    return src;
-  }
-}
-
 // ============================================
 // GLOBAL STATE
 // ============================================
 
-let lastFinalUrl = null;
 const clientRouterVersion = typeof window.__spa_router_version !== 'undefined' ? window.__spa_router_version : null;
 
 // ============================================
@@ -51,7 +56,19 @@ const clientRouterVersion = typeof window.__spa_router_version !== 'undefined' ?
 const SPA_CONFIG = {
   enabled: true,
   excludedRoutes: ['/delegation'],
-  scrollToTopOnNavigate: true
+  scrollToTopOnNavigate: true,
+  containerSelectorsCore: [
+    'nav#topbar-first',
+    'div#topbar-second',
+    'main'
+  ],
+  // Keep extended container syncing enabled for compatibility while refactoring.
+  // This can be disabled later after integration testing.
+  enableExtendedContainerSync: true,
+  extendedContainerSelectors: [
+    'aside',
+    'section'
+  ]
 };
 
 // ============================================
@@ -67,21 +84,33 @@ function isSPARoute(path) {
   return !SPA_CONFIG.excludedRoutes.some(route => path === route || path.startsWith(route + '/'));
 }
 
+const navigationAdapter = createNavigationAdapter({
+  isSPARoute,
+  loadContent
+});
+
+const domSwapPipeline = createDomSwapPipeline({
+  clientRouterVersion,
+  spaConfig: SPA_CONFIG,
+  cleanupTooltips,
+  classifyScript,
+  executeScripts,
+  reinitializeDynamicContent,
+  scrollToTopInstant
+});
+
+const contentLoader = createContentLoader({
+  showTimeoutModal,
+  replaceContainerContent
+});
+
 /**
  * Check if a link is an internal link (same origin)
  * @param {HTMLAnchorElement} link - The anchor element
  * @returns {boolean} True if internal link
  */
 function isInternalLink(link) {
-  const href = link.getAttribute('href');
-  if (!href) return false;
-  if (href.startsWith('#') || href.startsWith('javascript:') || href.startsWith('mailto:')) return false;
-  try {
-    const url = new URL(href, window.location.href);
-    return url.origin === window.location.origin;
-  } catch (e) {
-    return false;
-  }
+  return navigationAdapter.isInternalLink(link);
 }
 
 /**
@@ -89,56 +118,7 @@ function isInternalLink(link) {
  * @param {Event} e - Click event
  */
 function handleLinkClick(e) {
-  const link = e.target.closest('a');
-  if (!link) return;
-  
-  let href = link.getAttribute('href');
-  
-  // Ignore anchor-only links (e.g., href="#" or href="#back-to-top")
-  if (href && href.startsWith('#')) {
-    console.debug('[SPA Router] Click: Anchor link, allowing default behavior');
-    return;
-  }
-  
-  // Ignore links that are meant to open modals (e.g., modal-open class)
-  if (link.classList.contains('modal-open')) {
-    console.debug('[SPA Router] Click: Modal link, allowing default/modal behavior');
-    return;
-  }
-
-  // Ignore links with data-fancybox attribute (for Fancybox lightbox)
-  if (link.hasAttribute('data-fancybox')) {
-    console.debug('[SPA Router] Click: Fancybox link, pushing marker state and allowing default/fancybox behavior');
-    history.pushState({ __fancyboxMarker: true }, '', window.location.href);
-    return;
-  }
-
-  // Ignore links with inline event handlers (onclick, etc.)
-  if (link.hasAttribute("onclick") || link.onclick || link.hasAttribute("data-spa-ignore")) {
-    console.debug("[SPA Router] Click: Link has event handler or data-spa-ignore, allowing default behavior");
-    return;
-  }
-  
-  console.debug('[SPA Router] Click: Found anchor, href=', href);
-  
-  if (!isInternalLink(link)) {
-    console.debug('[SPA Router] Click: Not an internal link, skipping');
-    return;
-  }
-
-  if (!href.startsWith('http://') && !href.startsWith('https://') && !href.startsWith('/')) {
-    href = '/' + href;
-  }
-  
-  // Allow middle-click, ctrl-click, cmd-click to open in new tab
-  if (e.button !== 0 || e.ctrlKey || e.metaKey || e.shiftKey) {
-    console.debug('[SPA Router] Click: Modified click (middle/ctrl/cmd/shift), skipping');
-    return;
-  }
-
-  console.debug('[SPA Router] Click: Preventing default, navigating to', href);
-  e.preventDefault();
-  navigateTo(href);
+  navigationAdapter.handleLinkClick(e);
 }
 
 /**
@@ -171,31 +151,7 @@ function scrollToTopInstant() {
  * @param {string} url - The URL to navigate to
  */
 function navigateTo(url) {
-  console.debug('[SPA Router] Navigate: url=', url);
-  
-  const path = new URL(url, window.location.href).pathname;
-  console.debug('[SPA Router] Navigate: parsed path=', path);
-  
-  // Check if route is SPA-capable
-  if (!isSPARoute(path)) {
-    console.debug('[SPA Router] Navigate: Not an SPA route, falling back to full reload');
-    window.location.href = url;
-    return;
-  }
-  
-  console.debug('[SPA Router] Navigate: SPA route detected, pushing state and loading content');
-  
-  // Dispatch event to pause live updates before navigation
-  const beforeEvent = new CustomEvent('spa:beforeNavigate', {
-    detail: { path: path, url: url }
-  });
-  window.dispatchEvent(beforeEvent);
-  
-  // Update History API
-  history.pushState({ path, spa: true, __friendicaSPA: true }, '', url);
-
-  // Load content
-  loadContent(url);
+  navigationAdapter.navigateTo(url);
 }
 
 // ============================================
@@ -203,92 +159,12 @@ function navigateTo(url) {
 // ============================================
 
 /**
- * Show timeout modal overlay
- * Displays a modal dialog that can be clicked away
- */
-function showTimeoutModal() {
-  console.debug('[SPA Router] showTimeoutModal: Displaying timeout overlay');
-  
-  hideLoading();
-  
-  // Check if modal already exists
-  if (document.getElementById('spa-timeout-modal')) {
-    console.debug('[SPA Router] showTimeoutModal: Modal already exists');
-    return;
-  }
-  
-  // Get translated texts from PHP
-  const title = window.spaErrorTexts?.timeout || 'Timeout';
-  const message = window.spaErrorTexts?.timeout_message || 'Request timed out';
-  const closeText = window.spaErrorTexts?.close || 'Close';
-  
-  // Create modal overlay
-  const modal = document.createElement('div');
-  modal.id = 'spa-timeout-modal';
-  modal.className = 'spa-modal-overlay';
-  
-  // Create modal content
-  const content = document.createElement('div');
-  content.className = 'spa-modal-content';
-  
-  // Create heading
-  const heading = document.createElement('h2');
-  heading.textContent = title;
-  
-  // Create message paragraph
-  const messageElement = document.createElement('p');
-  messageElement.textContent = message;
-  
-  // Create close button
-  const closeButton = document.createElement('button');
-  closeButton.textContent = closeText;
-  closeButton.className = 'btn btn-primary spa-modal-close-btn';
-  closeButton.onclick = dismissTimeoutModal;
-  
-  content.append(heading, messageElement, closeButton);
-  modal.appendChild(content);
-  
-  // Close modal when clicking on overlay (outside content)
-  modal.addEventListener('click', (e) => { if (e.target === modal) dismissTimeoutModal(); });
-  
-  // Add modal to body
-  document.body.appendChild(modal);
-  
-  // Also listen for Escape key
-  const escapeHandler = (e) => { if (e.key === 'Escape' || e.keyCode === 27) dismissTimeoutModal(); };
-  document.addEventListener('keydown', escapeHandler);
-  
-  // Store reference to clean up
-  modal._escapeHandler = escapeHandler;
-}
-
-/**
- * Dismiss the timeout modal
- */
-function dismissTimeoutModal() {
-  const modal = document.getElementById('spa-timeout-modal');
-  if (modal) {
-    if (modal._escapeHandler) {
-      document.removeEventListener('keydown', modal._escapeHandler);
-    }
-    modal.remove();
-    console.debug('[SPA Router] dismissTimeoutModal: Modal removed');
-  }
-}
-
-/**
  * Handle HTTP error responses
  * @param {number} status - HTTP status code
  * @param {string} url - The original request URL
  */
 function handleHttpError(status, url) {
-  console.debug('[SPA Router] handleHttpError: status=', status, 'url=', url);
-  
-  // 401: redirect to login
-  if (status === 401) {
-    console.debug('[SPA Router] handleHttpError: 401 Unauthorized - redirecting to login');
-    window.location.href = '/login?return_path=' + encodeURIComponent(url);
-  }
+  contentLoader.handleHttpError(status, url);
 }
 
 /**
@@ -296,139 +172,12 @@ function handleHttpError(status, url) {
  * @param {string} url - The URL to load
  */
 function loadContent(url) {
-  console.debug('[SPA Router] LoadContent: url=', url);
-  
-  // Use normal URL without SPA parameter - server returns full HTML
-  const fetchUrl = new URL(url, window.location.href);
-  console.debug('[SPA Router] LoadContent: fetching URL=', fetchUrl.toString());
-  
-  // Track the final URL after all redirects
-  let finalUrl = fetchUrl.toString();
-
-  showFetching();
-  fetch(fetchUrl, { headers: { 'Accept': 'text/html' }, credentials: 'include' })
-    .then(async (response) => {
-      
-      console.debug('[SPA Router] LoadContent: Response received, status=', response.status, 'response.url=', response.url);
-      
-      // Check Content-Type header - only text/html is allowed for SPA
-      const contentType = response.headers.get('Content-Type') || response.headers.get('content-type') || '';
-      if (!contentType.includes('text/html')) {
-        console.debug('[SPA Router] LoadContent: Invalid Content-Type:', contentType, '- falling back to full reload');
-        window.location.href = url;
-        return null;
-      }
-      
-      // Get the final URL after any redirects - with automatic following, response.url contains it
-      if (response.url && response.url !== fetchUrl.toString()) {
-        finalUrl = response.url;
-        console.debug('[SPA Router] LoadContent: Final URL after redirects:', finalUrl);
-      }
-      
-      // Special handling for 401 and 504
-      if (response.status === 401) {
-        console.debug('[SPA Router] LoadContent: 401 Unauthorized - redirecting to login');
-        handleHttpError(401, url);
-        return null;
-      }
-      
-      if (response.status === 504) {
-        console.debug('[SPA Router] LoadContent: 504 Gateway Timeout - showing modal');
-        showTimeoutModal();
-        return null;
-      }
-      
-      // For all other responses (including other HTTP errors like 404, 500, etc.)
-      // load the server's error page content via SPA
-      showReceiving();
-      
-      const html = await response.text();
-      
-      console.debug('[SPA Router] LoadContent: HTML received, type:', typeof html, 'length:', html ? html.length : 0, 'finalUrl:', finalUrl);
-      
-      showProcessing();
-      
-      // Validate that html is a string
-      if (typeof html !== 'string') {
-        console.error('[SPA Router] LoadContent: html is not a string! type:', typeof html, 'value:', html);
-        throw new Error('Response body is not a string');
-      }
-      
-      // Update history with the final URL if there were redirects
-      if (finalUrl !== fetchUrl.toString()) {
-        console.debug('[SPA Router] LoadContent: Updating history to final URL:', finalUrl);
-        history.replaceState({ path: new URL(finalUrl).pathname, spa: true, __friendicaSPA: true }, '', finalUrl);
-      }
-      
-      // Replace content of the three main containers
-      // Pass the final URL to detect display pages after redirects
-      replaceContainerContent(html, finalUrl);
-      hideLoading();
-      
-      console.debug('[SPA Router] LoadContent: Process completed successfully');
-      return html;
-    })
-    .catch(error => {
-      hideLoading();
-      console.error('[SPA Router] Error loading content:', error);
-      console.error('[SPA Router] Error stack:', error.stack);
-      console.error('[SPA Router] Error name:', error.name);
-      console.error('[SPA Router] Error message:', error.message);
-      console.error('[SPA Router] Error details:', {
-        url: url,
-        fetchUrl: fetchUrl.toString(),
-        finalUrl: finalUrl,
-        error: error
-      });
-      
-      // Handle timeout errors with modal - only for server 504, not for client timeout
-      // Client timeout (AbortError) just dismisses the delay modal, does not show timeout modal
-      if (error.name === 'AbortError') {
-        console.debug('[SPA Router] LoadContent: AbortError (client timeout) - dismissing delay modal only');
-        // Don't show timeout modal for client timeout, only dismiss delay modal
-      } else if (error.name === 'TypeError' && error.message.includes('Failed to fetch')) {
-        // Network error - fall back to full reload
-        console.debug('[SPA Router] LoadContent: Network error - falling back to reload');
-        window.location.href = url;
-      } else {
-        // Fallback: Full page reload for other errors
-        console.debug('[SPA Router] Falling back to full page reload for URL:', url);
-        window.location.href = url;
-      }
-    });
+  contentLoader.loadContent(url);
 }
 
 // ============================================
 // CONTAINER CONTENT REPLACEMENT
 // ============================================
-
-/**
- * Remove all tooltip elements to prevent ghost tooltips after SPA navigation
- * Targets common tooltip classes and dynamically created tooltip elements
- */
-function cleanupTooltips() {
-  const tooltipSelectors = [
-    '.tooltip',                    // Standard tooltip class
-    '[class*="tooltip"]',          // Classes containing "tooltip"
-    '.ui-tooltip',                // jQuery UI tooltips
-    '.popover',                   // Bootstrap popovers
-    '[role="tooltip"]',           // ARIA role tooltips
-    '.fancybox-wrap',             // Fancybox
-    '.colorbox',                  // Colorbox
-    '.jGrowl',                    // jGrowl notifications
-    '.panel'                      // Friendica permission panel (from lockview)
-  ];
-
-  tooltipSelectors.forEach(selector => {
-    const elements = document.querySelectorAll(selector);
-    // Convert NodeList to Array for compatibility with older browsers
-    const elementsArray = Array.prototype.slice.call(elements);
-    elementsArray.forEach(el => {
-      console.debug('[SPA Router] Cleanup: Removing tooltip element:', selector);
-      el.remove();
-    });
-  });
-}
 
 /**
  * Replace content of the main containers: nav#topbar-first, div#topbar-second, and main
@@ -437,374 +186,7 @@ function cleanupTooltips() {
  * @param {string} finalUrl - The final URL after following redirects (optional)
  */
 function replaceContainerContent(htmlString, finalUrl = null) {
-  console.debug('[SPA Router] ReplaceContent: starting replacement');
-  
-  // Clean up any existing tooltips first to prevent ghost elements
-  cleanupTooltips();
-  
-  // Validate htmlString is a string
-  if (typeof htmlString !== 'string') {
-    console.error('[SPA Router] ReplaceContent: htmlString is not a string!');
-    return;
-  }
-  
-  // Store the final URL for potential redirect handling
-  lastFinalUrl = finalUrl || window.location.href;
-
-  // Use DOMParser for safer and more robust parsing
-  const parser = new DOMParser();
-  const newDoc = parser.parseFromString(htmlString, 'text/html');
-  
-  // 1. Update document title
-  const newTitle = newDoc.querySelector('title');
-  if (newTitle) document.title = newTitle.textContent;
-
-  // Check for router version update in the new content
-  const newRouterVersion = newDoc.querySelector('script[data-spa-version]');
-  if (newRouterVersion) {
-    const serverVersion = newRouterVersion.getAttribute('data-spa-version');
-    if (clientRouterVersion && serverVersion && serverVersion !== clientRouterVersion) {
-      console.debug('[SPA Router] Version mismatch detected (Client:', clientRouterVersion, 'Server:', serverVersion, ') - triggering full reload');
-      window.location.reload();
-      return;
-    }
-  }
-
-  // 2. Identify all resources from head and body
-  const globalScripts = [];
-  const bodyScripts = [];
-  const externalScriptPromises = [];
-  
-  // Collect all script and link/style elements from the whole response
-  const resourceElements = newDoc.querySelectorAll('head > link, head > script, head > style, body script, body link');
-  
-  // 3. Define the main containers to update with theme-agnostic fallbacks.
-  const containerSelectors = [
-    'nav#topbar-first',
-    'div#topbar-second',
-    'main',
-    'aside',
-    'right_aside',
-    'section'
-  ];
-  
-  // 4. PERFORM CONTENT REPLACEMENT (Main body)
-  containerSelectors.forEach(selector => {
-    const newDiv = newDoc.querySelector(selector);
-    const oldDiv = document.querySelector(selector);
-    
-    if (newDiv && oldDiv) {
-      console.debug('[SPA Router] ReplaceContent: Updating container', selector);
-      
-      // Extract scripts within this container to handle manually.
-      // We pass both globalScripts and bodyScripts array to ensure that
-      // scripts like infinite_scroll (which contain definitions) are 
-      // correctly classified as global headers even if they happen to 
-      // be located inside a body container.
-      const inlineScripts = newDiv.querySelectorAll('script:not([src])');
-      for (const script of inlineScripts) {
-        const content = script.textContent.trim();
-        if (content) classifyScript(content, globalScripts, bodyScripts, script, 'container');
-        script.parentNode.removeChild(script);
-      }
-
-      // Special handling for scrolling on main container
-      if (selector === 'main') {
-        const effectivePath = lastFinalUrl ? new URL(lastFinalUrl).pathname : window.location.pathname;
-        if (SPA_CONFIG.scrollToTopOnNavigate && !effectivePath.includes('/display/')) {
-          scrollToTopInstant();
-        }
-      }
-      
-      oldDiv.innerHTML = newDiv.innerHTML;
-    }
-  });
-
-  // Handle scroll-loader
-  const scrollLoaderNew = newDoc.querySelector('#scroll-loader');
-  if (scrollLoaderNew) {
-    const scrollLoaderOld = document.getElementById('scroll-loader');
-    if (scrollLoaderOld) scrollLoaderOld.innerHTML = scrollLoaderNew.innerHTML;
-  }
-
-  // Collect all external scripts from the new page that we need to keep
-  const newPageScripts = new Set();
-  const newPageResourceElements = newDoc.querySelectorAll('head > script[src], body script[src]');
-  newPageResourceElements.forEach(el => {
-    const src = normalizeScriptSource(el.getAttribute('src'));
-    if (src) {
-      newPageScripts.add(src);
-    }
-  });
-
-  // Remove old Subscribers
-  if (typeof window.clearSPASubscribers === 'function') {
-      const cleanupStats = window.clearSPASubscribers(newPageScripts);
-      console.debug('[SPA Router] Subscriber cleanup stats:', cleanupStats);
-  }
-
-  // Remove scripts from current DOM that are NOT in the new page and NOT persistent
-  const currentScripts = document.querySelectorAll('script[src]');
-  console.debug('[SPA Router] Found', currentScripts.length, 'external script(s) in current DOM');
-  console.debug('[SPA Router] New page has', newPageScripts.size, 'external script(s)');
-  currentScripts.forEach(script => {
-    const src = normalizeScriptSource(script.getAttribute('src') || script.src);
-    if (newPageScripts.has(src) || script.hasAttribute('data-spa-persistent')) {
-      console.debug('[SPA Router] Keeping script (in new page or persistent):', src);
-      return;
-    }
-
-    console.debug('[SPA Router] Removing script (not in new page, not persistent):', src);
-    script.parentNode.removeChild(script);
-  });
-
-  // 5. SYNC HEAD (Links, Styles, External Scripts)
-
-  for (const el of resourceElements) {
-    const tag = el.tagName.toLowerCase();
-    
-    if (tag === 'link' || tag === 'style') {
-      const href = el.getAttribute('href');
-      const selector = tag + (href ? `[href="${href}"]` : '');
-      if (!document.head.querySelector(selector)) {
-        console.debug('[SPA Router] Adding style/link:', selector);
-        document.head.appendChild(el.cloneNode(true));
-      }
-    } else if (tag === 'script') {
-      const src = el.getAttribute('src');
-      if (src) {
-        const normalizedSrc = normalizeScriptSource(src);
-        // Only add if not already present in DOM
-        if (!document.querySelector(`script[src="${src}"]`) && !document.querySelector(`script[src="${normalizedSrc}"]`)) {
-          console.debug('[SPA Router] Adding script (not in DOM):', src);
-          const clone = document.createElement('script');
-          
-          // Copy all attributes
-          Array.from(el.attributes).forEach(attr => {
-            clone.setAttribute(attr.name, attr.value);
-          });
-          
-          // Ensure scripts are executed in order
-          clone.async = false;
-          
-          const p = new Promise(resolve => {
-            clone.onload = () => {
-              console.debug('[SPA Router] Script loaded:', src);
-              resolve();
-            };
-            clone.onerror = () => {
-              console.warn('[SPA Router] Script load error:', src);
-              resolve(); 
-            };
-            // Do NOT use a safety timeout here if we want to ensure order
-          });
-          externalScriptPromises.push(p);
-          document.head.appendChild(clone);
-        } else {
-          console.debug('[SPA Router] Script already in DOM, skipping add:', normalizedSrc || src);
-        }
-      } else {
-        // It's an inline script from head (or already processed if from body)
-        // Only classify if it hasn't been removed from content already
-        if (el.parentNode && el.parentNode.tagName.toLowerCase() === 'head') {
-          classifyScript(el.textContent.trim(), globalScripts, bodyScripts, el, 'head');
-        }
-      }
-    }
-  }
-
-  // 6. EXECUTION LIFECYCLE
-  window.__spa_executing_page_scripts = true;
-  window.__spa_executing_fragment_scripts = false;
-  window.__spa_bodyScripts = bodyScripts;
-
-  // Execute global-like headers first (vars, aStr, etc.)
-  window.__spa_executing_fragment_scripts = true;
-  executeScripts(globalScripts, 'global-head');
-  
-  // IMMEDIATELY re-initialize infinite scroll if the new page provides the config
-  // (the 'infinite_scroll' object is usually in the global-head or body-scripts)
-  if (typeof initInfiniteScroll === 'function') {
-    initInfiniteScroll();
-  }
-  
-  // Also dispatch the custom event for any other listeners
-  window.dispatchEvent(new CustomEvent('spa:initInfiniteScroll'));
-  
-  window.__spa_executing_fragment_scripts = false;
-
-  const runFinalAppInit = () => {
-    window.__spa_reinit_phase = true;
-    reinitializeDynamicContent();
-    window.__spa_reinit_phase = false;
-    window.__spa_executing_page_scripts = false;
-  };
-  if (externalScriptPromises.length > 0) {
-    console.debug('[SPA Router] Waiting for', externalScriptPromises.length, 'scripts before reinit...');
-    Promise.all(externalScriptPromises).then(() => {
-      // Small timeout to ensure the browser has a chance to execute the code
-      // of the newly loaded scripts before we trigger the reinitialization.
-      setTimeout(runFinalAppInit, 10);
-    });
-  } else {
-    runFinalAppInit();
-  }
-}
-
-// ============================================
-// SCRIPT PROCESSING
-// ============================================
-
-/**
- * Check AST for DOM-ready patterns (jQuery ready, DOMContentLoaded, load events)
- * @param {Object} ast - The parsed AST
- * @returns {boolean} - True if DOM-ready handler detected
- */
-function checkForDOMReadyPatterns(ast) {
-  function walk(node) {
-    if (!node || typeof node !== 'object') return false;
-    if (node.type === 'CallExpression' && node.callee.type === 'Identifier') {
-      if (['$', 'jQuery'].includes(node.callee.name) && node.arguments.length > 0) {
-        if (node.arguments[0].type === 'FunctionExpression') return true;
-      }
-    }
-    if (node.type === 'CallExpression' && node.callee.type === 'MemberExpression') {
-      if (node.callee.property.name === 'ready') return true;
-      if (node.callee.property.name === 'addEventListener' && node.callee.object.type === 'Identifier' && ['document', 'window'].includes(node.callee.object.name)) {
-        const eventArg = node.arguments[0];
-        if (eventArg?.type === 'Literal' && ['DOMContentLoaded', 'load'].includes(eventArg.value)) return true;
-      }
-    }
-    for (const key in node) {
-      if (key === 'parent' || key === 'start' || key === 'end' || key === 'loc' || key === 'range') continue;
-      const child = node[key];
-      if (Array.isArray(child)) {
-        for (const c of child) if (walk(c)) return true;
-      } else if (walk(child)) return true;
-    }
-    return false;
-  }
-  return walk(ast);
-}
-
-/**
- * Classify script content using AST analysis
- * @param {string} content - The script content
- * @param {string} source - Source context ('head', 'container', etc.)
- * @returns {'global'|'body'|null}
- */
-function classifyScriptContent(content, source) {
-  const ast = acorn.parse(content, { ecmaVersion: 2020, sourceType: 'script' });
-  if (checkForDOMReadyPatterns(ast)) return 'body';
-  if (ast.body.length > 0) {
-    const firstNode = ast.body[0];
-    if (firstNode.type === 'VariableDeclaration') return 'global';
-    if (firstNode.type === 'FunctionDeclaration') return 'global';
-    if (firstNode.type === 'ClassDeclaration') return 'global';
-    if (content.trim().startsWith('aStr')) return 'global';
-  }
-  return null;
-}
-
-/**
- * Transform code to promote top-level declarations to window scope
- * @param {string} code - The JavaScript code to transform
- * @returns {string} - Transformed code
- */
-function promoteToGlobal(code) {
-  const ast = acorn.parse(code, { ecmaVersion: 2020, sourceType: 'script', locations: true, ranges: true });
-  let newCode = '';
-  let lastEnd = 0;
-  for (const node of ast.body) {
-    if (node.start > lastEnd) newCode += code.slice(lastEnd, node.start);
-    if (node.type === 'VariableDeclaration') {
-      for (const declaration of node.declarations) {
-        if (declaration.id.type === 'Identifier') {
-          const varName = declaration.id.name;
-          const initCode = declaration.init ? code.slice(declaration.init.start, declaration.init.end) : 'undefined';
-          newCode += `window.${varName} = ${initCode};`;
-        } else {
-          newCode += code.slice(node.start, node.end);
-        }
-      }
-    }
-    else if (node.type === 'FunctionDeclaration') {
-      newCode += `window.${node.id.name} = ${code.slice(node.start, node.end)};`;
-    }
-    else if (node.type === 'ClassDeclaration') {
-      newCode += `window.${node.id.name} = ${code.slice(node.start, node.end)};`;
-    }
-    else {
-      newCode += code.slice(node.start, node.end);
-    }
-    lastEnd = node.end;
-  }
-  if (lastEnd < code.length) newCode += code.slice(lastEnd);
-  return newCode;
-}
-
-/**
- * Classify a script as either global (definitions) or body (execution)
- * @param {string} content
- * @param {Array} globalScripts
- * @param {Array} bodyScripts
- * @param {HTMLScriptElement|null} scriptEl
- * @param {string} source
- */
-function classifyScript(content, globalScripts, bodyScripts, scriptEl = null, source = '') {
-  if (!content) return;
-  if (scriptEl) {
-    const attr = (scriptEl.getAttribute('data-spa-scope') || scriptEl.getAttribute('data-spa-script') || '').toLowerCase();
-    if (attr === 'global' || attr === 'head') { globalScripts.push(content); return; }
-    if (attr === 'body' || attr === 'fragment') { bodyScripts.push(content); return; }
-  }
-  const classification = classifyScriptContent(content, source);
-  if (classification === 'body') { bodyScripts.push(content); return; }
-  if (classification === 'global') { globalScripts.push(content); return; }
-  if (source === 'head') { globalScripts.push(content); return; }
-  bodyScripts.push(content);
-}
-
-/**
- * Execute a list of script contents
- * @param {Array} scripts
- * @param {string} context
- */
-function executeScripts(scripts, context) {
-  if (!scripts || scripts.length === 0) return;
-  
-  console.debug('[SPA Router] executeScripts: Executing ' + scripts.length + ' ' + context + ' scripts');
-  
-  // We combine all scripts of the same category (global or body) into a single 
-  // execution block. This ensures they share a lexical scope, allowing one 
-  // script to use 'const' or 'let' variables defined in another script 
-  // from the same page load.
-  const combinedContent = scripts.join('\n\n/* --- Next Script --- */\n\n');
-  try {
-    // 1. Lexical Scoping:
-    // We wrap the combined scripts in an anonymous block { ... } to prevent 
-    // "redeclaration" errors (TypeError: redeclaration of const) when 
-    // navigating between pages in SPA mode. This block creates a new 
-    // lexical scope for each page load.
-    
-    // 2. Variable & Function promotion to global scope:
-    // Use AST parser for robust transformation
-    const promotedContent = promoteToGlobal(combinedContent);
-
-    // 3. Runtime Safety: 
-    // We wrap the whole block in a try-catch. This is important because 
-    // assignments to global variables that were previously defined as 'const' 
-    // during the initial page load will throw a TypeError. We want to catch 
-    // these gracefully so other scripts can continue.
-    const wrappedContent = 'try {\n{\n' + promotedContent + '\n}\n} catch (e) { console.debug("[SPA Router] Handled error in scripts:", e.message); }';
-
-    const scriptEl = document.createElement('script');
-    scriptEl.textContent = wrappedContent;
-    document.head.appendChild(scriptEl);
-    document.head.removeChild(scriptEl);
-  } catch (e) {
-    console.error('[SPA Router] Error executing combined ' + context + ' scripts:', e);
-  }
+  domSwapPipeline.replaceContainerContent(htmlString, finalUrl);
 }
 
 // ============================================
@@ -812,81 +194,11 @@ function executeScripts(scripts, context) {
 // ============================================
 
 /**
- * Trigger spa:document:ready event
- */
-function triggerSPADocumentReady() {
-  window.dispatchEvent(new CustomEvent('spa:document:ready'));
-}
-
-/**
- * Trigger spa:window:load event
- */
-function triggerSPAWindowLoad() {
-  console.debug('[SPA Router] triggerSPAWindowLoad: dispatching spa:window:load');
-  window.dispatchEvent(new CustomEvent('spa:window:load'));
-}
-
-/**
- * Re-initialize dynamic content after SPA navigation
- * This is important for elements that need event listeners
- */
-function reinitializeDynamicContent() {
-  // Execute body scripts (including widget init scripts) after DOM insertion
-  if (window.__spa_bodyScripts?.length > 0) {
-    console.debug('[SPA Router] reinitializeDynamicContent: Executing ' + window.__spa_bodyScripts.length + ' body scripts');
-    window.__spa_executing_fragment_scripts = true;
-    executeScripts(window.__spa_bodyScripts, 'body-scripts');
-    window.__spa_executing_fragment_scripts = false;
-    // Clear the stored scripts after execution
-    window.__spa_bodyScripts = [];
-  }
-
-  // Dispatch custom events for other scripts to hook into
-  const spaNavigateEvent = new CustomEvent('spa:navigate', {
-    detail: { path: window.location.pathname }
-  });
-  window.dispatchEvent(spaNavigateEvent);
-  
-  // Trigger window loaded event for elements that need all resources (images, etc.)
-  console.debug('[SPA Router] reinitializeDynamicContent: dispatching spa:window:load');
-  const spaDocumentLoadEvent = new CustomEvent('spa:window:load');
-  window.dispatchEvent(spaDocumentLoadEvent);
-  
-  // Trigger DOM ready for elements that need to reinitialize after SPA navigation
-  const spaDocumentReadyEvent = new CustomEvent('spa:document:ready');
-  window.dispatchEvent(spaDocumentReadyEvent);
-  
-  // Trigger infinite scroll reinitialization for network pages
-  const initInfiniteScrollEvent = new CustomEvent('spa:initInfiniteScroll');
-  window.dispatchEvent(initInfiniteScrollEvent);
-  
-  // Trigger NavUpdate to check for unread posts immediately after SPA navigation
-  if (typeof NavUpdate === 'function') {
-    console.debug('[SPA Router] Calling NavUpdate after navigation');
-    NavUpdate();
-  }
-}
-
-/**
  * Handle browser back/forward navigation
  * @param {Event} e - Popstate event
  */
 function handlePopState(e) {
-  console.debug('[SPA Router] PopState: state=', e.state, 'hash=', window.location.hash);
-  
-  if (e.state?.spa && e.state.__friendicaSPA) {
-    console.debug('[SPA Router] PopState: SPA navigation detected, loading', window.location.href);
-    
-    // Dispatch event to pause live updates before popstate navigation
-    const beforeEvent = new CustomEvent('spa:beforeNavigate', {
-      detail: { path: window.location.pathname, url: window.location.href }
-    });
-    window.dispatchEvent(beforeEvent);
-    
-    loadContent(window.location.href);
-  } else {
-    console.debug('[SPA Router] PopState: Not an SPA state or not from our router, ignoring');
-  }
+  navigationAdapter.handlePopState(e);
 }
 
 // ============================================
@@ -904,11 +216,7 @@ function initSPARouter() {
     return;
   }
   
-  // Intercept link clicks (using event delegation)
-  document.addEventListener('click', handleLinkClick);
-  
-  // Browser navigation (back/forward)
-  window.addEventListener('popstate', handlePopState);
+  navigationAdapter.bindNavigationEvents();
   
   // Register handlers for jQuery ready and window load events
   if (typeof $ !== 'undefined') {
